requirements.txt
arxiv
chromadb
deepl
llamaapi
langchain
langchain-experimental
langsmith
PyMuPDF
python-telegram-bot
pytz
PyYAML
regex
requests
selenium
tokenizers
unstructured
urllib3
